Make mpi4py an optional dependency.
